JSONRPC client WIP
